PWN-5267 - RpcSolanaApi and ren pool base url fixes
