+++ app/src/main/java/org/p2p/wallet/root/RootActivity.kt
@@ -6,7 +6,6 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 import com.google.android.material.snackbar.Snackbar
@@ -59,7 +58,6 @@ class RootActivity : BaseMvpActivity<RootContract.View, RootContract.Presenter>(
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(R.style.WalletTheme)
         super.onCreate(savedInstanceState)
-        val splashScreen = installSplashScreen()
         binding = ActivityRootBinding.inflate(LayoutInflater.from(this))
         setContentView(binding.root)
 
@@ -141,7 +139,7 @@ class RootActivity : BaseMvpActivity<RootContract.View, RootContract.Presenter>(
                 when (state) {
                     is SolanaNetworkState.Offline -> showSnackbarErrorMessage()
                     is SolanaNetworkState.Online,
-                    is SolanaNetworkState.Idle -> hideSnackbar()
+                    is SolanaNetworkState.Idle -> Unit
                 }
             }
         }
@@ -158,10 +156,5 @@ class RootActivity : BaseMvpActivity<RootContract.View, RootContract.Presenter>(
                 networkObserver.setSnackbarHidden(isHidden = true)
             }
         )
-    }
-
-    private fun hideSnackbar() {
-        snackbar?.dismiss()
-        snackbar = null
     }
 }
