+++ solanaj/src/main/java/org/p2p/solanaj/rpc/RpcSolanaRepository.kt
@@ -3,25 +3,11 @@ package org.p2p.solanaj.rpc
 import org.p2p.solanaj.core.Account
 import org.p2p.solanaj.core.PublicKey
 import org.p2p.solanaj.core.Transaction
-import org.p2p.solanaj.kits.renBridge.renVM.types.ParamsSubmitMint
-import org.p2p.solanaj.kits.renBridge.renVM.types.ResponseQueryBlockState
-import org.p2p.solanaj.kits.renBridge.renVM.types.ResponseQueryConfig
-import org.p2p.solanaj.kits.renBridge.renVM.types.ResponseQueryTxMint
-import org.p2p.solanaj.kits.renBridge.renVM.types.ResponseSubmitTxMint
 import org.p2p.solanaj.model.types.AccountInfo
 import org.p2p.solanaj.model.types.SignatureInformationResponse
 import org.p2p.solanaj.rpc.model.RecentPerformanceSample
 
 interface RpcSolanaRepository {
-    suspend fun getQueryMint(txHash: String): ResponseQueryTxMint
-    suspend fun getQueryBlockState(): ResponseQueryBlockState
-    suspend fun getQueryConfig(): ResponseQueryConfig
-    suspend fun submitTx(
-        hash: String,
-        mintTx: ParamsSubmitMint.MintTransactionInput,
-        selector: String
-    ): ResponseSubmitTxMint
-
     suspend fun getAccountInfo(stateKey: PublicKey): AccountInfo
     suspend fun getRecentPerformanceSamples(exampleCount: Int): List<RecentPerformanceSample>
     suspend fun sendTransaction(transaction: Transaction, signer: Account): String
